feat: Reinitialize Course Module tabs on publish
Previously, it would only do it once and never try to
reinitialize the tabs again. This leads to an issue if you ever want
to update that list. Now we will check on every publish
and if we detect a change, will update the list of tabs.

This change also switches Dates Tab to be a default static tab.